feat(request-password-recovery): implement password recovery logic
- Add custom hook for password recovery logic
- Refactor RequestPasswordRecoveryScreen to use the new hook
- Update imports for schema and types
- Remove unused utils and index files
