+++ src/features/request-password-recovery/types.ts
@@ -1,4 +1,4 @@
 import * as z from 'zod/mini';
-import { RequestPasswordRecoverySchema } from './utils';
+import { RequestPasswordRecoverySchema } from './utils/schema';
 
 export type RequestPasswordRecoverySchemaType = z.infer<typeof RequestPasswordRecoverySchema>;
