--- src/features/request-password-recovery/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './schema';
-export * from './errors-case';
